+++ nuget-client/BrevgeneratorClient.Tests/GenererBrevArgsBuilderDefaultTemplateTests.cs
@@ -1,4 +1,4 @@
-using Arbeidstilsynet.Brevgenerator.Client.Model;
+using Arbeidstilsynet.Brevgenerator.Client.Models;
 using Xunit;
 
 namespace Arbeidstilsynet.Brevgenerator.Client.Tests;