+++ nuget-client/BrevgeneratorClient/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Publiserer nå til `nuget.org` i stedet for Azure DevOps
 - Omdøpt pakke fra `AT.Brevgenerator.Klient` til `Arbeidstilsynet.Brevgenerator.Client`
-- Omdøpt namespace fra `AT.Brevgenerator.Klient` / `AT.Brevgenerator.Klient.Model` til `Arbeidstilsynet.Brevgenerator.Client` / `Arbeidstilsynet.Brevgenerator.Client.Model`
+- Omdøpt namespace fra `AT.Brevgenerator.Klient` / `AT.Brevgenerator.Klient.Model` til `Arbeidstilsynet.Brevgenerator.Client` / `Arbeidstilsynet.Brevgenerator.Client.Models`
 - Omdøpt `BrevgeneratorKlient` til `BrevgeneratorClient`
 
 ## 3.3.0